+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.2.0
+
+* Adds a verbose flag. Informs about plugin features such as "provides commands" or "action hook" etc.
+* Adds a `--no-heads` flag. Swicth off the column descriptions.
+* Revamped the internal UI component to use a nifty DSL to describe the column layout.
+
 # v0.1.0
 
 * simple list of all loaded plugin names